metadata-transformation-service/pyproject.toml
[project]
name = "metadata-transformation-service"
title = "Metadata Transformer Service"
version = "0.1.7-dev"
description = "The transformer service converts metadata from one format to another, ensuring compatibility with various repository standards and requirements. It plays a crucial role in facilitating data interoperability and seamless content migration between repositories. By leveraging XSLT-based transformation technology, the transformer service efficiently converts metadata into multiple formats, enabling seamless integration with diverse repository systems."
authors = [
    {name = "Eko Indarto", email = "eko.indarto@dans.knaw.nl"}
]
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "akmi-utils>=0.1.6.2",
    "boltons>=25.0.0",
    "dynaconf>=3.2.10",
    "fastapi-utils>=0.8.0",
    "jinja2>=3.1.5",
    "jproperties>=2.1.2",
    "rdflib>=7.1.3",
    "requests>=2.32.3",
    "saxonche>=12.5.0",
    "typing-inspect>=0.9.0",
    "xmltodict>=0.14.2",
]
